==> scripts/01_BRMS_models.Rmd <==
---
title: "BRMS univariate models"
author: "Sam Straus and Melissa Guzman"
date: "12/04/2023"
output: html_document
---

```{r load packages and data}
library(brms)
library(stringr)
library(dplyr)
library(rstan)
library(ape)

### load traits
traits_phylo <- read.csv("../data/Traits_final_cleaned.csv", fileEncoding="latin1")
one_phylo <- readRDS("../data/final_phylo.rds")


A <- ape::vcv.phylo(one_phylo)
```
Null models, univariate, with year

```{r Null-univariate-withyear}
bf_dispersal_null_withyear <- bf(dispersal_km ~ log(Mass_kg) + dispersal_year) + lognormal(link = "identity")
bf_home_range_null_withyear <- bf(hr.radius ~ log(Mass_kg) + foraging_year) + lognormal(link = "identity")
bf_migration_null_withyear <- bf(Migration_km ~ log(Mass_kg) + migration_year) + hurdle_lognormal(link = 'identity')


null.disp_withyear <- brm(bf_dispersal_null_withyear, data = traits_phylo, iter = 3000)
null.for_withyear <- brm(bf_home_range_null_withyear, data = traits_phylo, iter = 3000)
null.mig_withyear <- brm(bf_migration_null_withyear, data = traits_phylo, iter = 3000)

## adding max vs mean as random effect for dispersal model
bf_dispersal_null_withyear_withmax <- bf(dispersal_km ~ log(Mass_kg) + dispersal_year + (log(Mass_kg)|dispersal_type)) + lognormal(link = "identity")


null.disp_withyear_withmax <- brm(bf_dispersal_null_withyear_withmax, data = traits_phylo,
                      control = list(adapt_delta = 0.995, max_treedepth=20), iter = 5000)




```


save univariate null mods
```{r save-uni-nulls}
saveRDS(null.disp_withyear, "../mods/null.disp_withyear.rds")
saveRDS(null.for_withyear, "../mods/null.for_withyear.rds")
saveRDS(null.mig_withyear, "../mods/null.mig_withyear.rds") 

saveRDS(null.disp_withyear_withmax, "../mods/null.disp_withyear_withmax.rds")


# read mods
null.disp_withyear <- readRDS("../mods/null.disp_withyear.rds")
null.for_withyear <- readRDS("../mods/null.for_withyear.rds")
null.mig_withyear <- readRDS("../mods/null.mig_withyear.rds")

fixef(null.disp_withyear)
fixef(null.disp_withyear_withmax)
ranef(null.disp_withyear_withmax)


```
+/-


Null Model, multivariate, with year
```{r Null multivar with year}
bf_dispersal_null <- bf(dispersal_km ~ log(Mass_kg) + dispersal_year) + 
  lognormal(link = "identity")
bf_home_range_null <- bf(hr.radius ~ log(Mass_kg) + foraging_year) + 
  lognormal(link = "identity")
bf_migration_null <- bf(Migration_km ~ log(Mass_kg) + migration_year) +
  hurdle_lognormal(link = "identity")


null_multivar <- brm(bf_dispersal_null + bf_home_range_null + bf_migration_null, 
                     data = traits_phylo, cores = 4, iter = 3000)

saveRDS(null_multivar, "../mods/null_multivar.rds")

kable(class_type)
```


Trophic guild, random slope and intercept, include year

```{r guild-univariate-year}
bf_dispersal_trophic_withyear <- bf(dispersal_km ~ log(Mass_kg) + (log(Mass_kg)|diet_broadest_cat) + dispersal_year) + lognormal(link = "identity")
bf_home_range_trophic_withyear <- bf(hr.radius ~ log(Mass_kg)+ (log(Mass_kg)|diet_broadest_cat) + foraging_year) + lognormal(link = "identity")
bf_migration_trophic_withyear <- bf(Migration_km ~ log(Mass_kg) + (log(Mass_kg)|diet_broadest_cat) + migration_year) + hurdle_lognormal(link = 'identity')


trophic_year_prior_disp <- prior(normal(0,5),class = 'b', coef = logMass_kg) +
  prior(normal(0,5),class = 'b', coef = dispersal_year) +
  prior(student_t(5, 0, 3), class = 'Intercept') +
  prior(student_t(6, 0, 2.9), class = 'sd') +
  prior(student_t(5, 0, 2.9), class = 'sigma')

## works
trophic.disp_withyear <- brm(bf_dispersal_trophic_withyear, data = traits_phylo, 
                          control = list(adapt_delta = 0.995, max_treedepth=20), 
                          iter = 5000, cores = 16,
                          prior = trophic_year_prior_disp)

trophic_year_prior_for <- prior(normal(0,5),class = 'b', coef = logMass_kg) +
  prior(normal(0,5),class = 'b', coef = foraging_year) +
  prior(student_t(5, 0, 3), class = 'Intercept') +
  prior(student_t(6, 0, 2.9), class = 'sd') +
  prior(student_t(5, 0, 2.9), class = 'sigma')

# works
trophic.for_withyear <- brm(bf_home_range_trophic_withyear, data = traits_phylo, 
                         control = list(adapt_delta = 0.995, max_treedepth=20), iter = 5000, cores = 16,
                         prior = trophic_year_prior_for)


## works
trophic_year_prior_mig <- prior(normal(0,3),class = 'b', coef = logMass_kg) +
  prior(normal(0,3),class = 'b', coef = migration_year) +
  prior(student_t(5, 0, 3), class = 'Intercept') +
  prior(student_t(4, 0, 2.9), class = 'sd') +
  prior(student_t(4, 0, 2.9), class = 'sigma') 


trophic.mig_withyear <- brm(bf_migration_trophic_withyear, data = traits_phylo, 
                         control = list(adapt_delta = 0.995, max_treedepth=20), iter = 4000, cores = 16,
                         prior = trophic_year_prior_mig)


```

```{r save-guild-mods}

saveRDS(trophic.disp_withyear, "../mods/trophic.disp_withyear.rds")
saveRDS(trophic.for_withyear, "../mods/trophic.for_withyear.rds")
saveRDS(trophic.mig_withyear, "../mods/trophic.mig_withyear.rds") 

```


Trophic guild, mulivariate with year
```{r trophic-multivar-year}

trophic_withyear_prior_multivar <- prior(normal(0,5), class = 'b', 
                              coef = logMass_kg, resp = "dispersalkm") +
  prior(normal(0,5),class = 'b', coef = dispersal_year, resp = "dispersalkm") +
  prior(student_t(6, 0, 2.9), class = 'sd', resp = "dispersalkm") +
  prior(student_t(5, 0, 2.9), class = 'sigma', resp = "dispersalkm")+
  prior(normal(0,5), class = 'b', coef = logMass_kg, resp = "hrradius") +
  prior(normal(0,5),class = 'b', coef = foraging_year, resp = 'hrradius') +
  prior(student_t(6, 0, 2.9), class = 'sd', resp = "hrradius") +
  prior(student_t(5, 0, 2.9), class = 'sigma', resp = "hrradius") +
  prior(normal(0,7), class = 'b', coef = logMass_kg, resp = "Migrationkm") +
  prior(normal(0,7),class = 'b', coef = migration_year, resp = "Migrationkm") +
  prior(student_t(5, 0, 3), class = 'Intercept', resp = "Migrationkm") +
  prior(student_t(4, 0, 2.9), class = 'sd', resp = "Migrationkm") +
  prior(student_t(4, 0, 2.9), class = 'sigma',resp = "Migrationkm")

trophic_multivar_withyear <- brm(bf_dispersal_trophic_withyear + bf_home_range_trophic_withyear +
    bf_migration_trophic_withyear, data = traits_phylo, 
    control = list(adapt_delta = 0.995, max_treedepth = 20), 
    iter = 7000, cores = 16, prior = trophic_withyear_prior_multivar)


 saveRDS(trophic_multivar_withyear, "../mods/trophic_multivar_withyear.rds")


```


Class, Univariate, Random intercept and slope, with year
```{r univariate-class-withyear}

# random slope, with year
bf_dispersal_class_withyear <- bf(dispersal_km ~ log(Mass_kg) + 
                      (log(Mass_kg)|p|class) + dispersal_year) + lognormal()
bf_home_range_class_withyear <- bf(hr.radius ~ log(Mass_kg)+ 
                      (log(Mass_kg)|p|class) + foraging_year) +  lognormal()
bf_migration_class_withyear <- bf(Migration_km ~ log(Mass_kg) + 
                      (log(Mass_kg)|p|class) + migration_year) + hurdle_lognormal()

class_withyear_prior <- prior(normal(0,5),class = 'b', coef = logMass_kg) +
  prior(student_t(4, 0, 4), class = 'Intercept') +
  prior(student_t(6, 0, 2.9), class = 'sd') +
  prior(student_t(4, 0, 2.9), class = 'sigma')

class.disp_withyear <- brm(bf_dispersal_class_withyear, data = traits_phylo, 
                         control = list(adapt_delta = 0.995, max_treedepth=20), 
                         iter = 3000, cores = 4, prior = class_withyear_prior)


class.for_withyear <- brm(bf_home_range_class_withyear, data = traits_phylo, 
                        control = list(adapt_delta = 0.995, max_treedepth=20), 
                        iter = 3000, cores = 4, prior = class_withyear_prior)



class.mig_withyear <- brm(bf_migration_class_withyear, data = traits_phylo, 
                        control = list(adapt_delta = 0.995, max_treedepth=20), 
                        iter = 3000, cores = 4, prior = class_withyear_prior)

```
Save rds files
```{r}
saveRDS(class.disp_withyear, "../mods/class.disp_withyear.rds")
saveRDS(class.for_withyear, "../mods/class.for_withyear.rds")
saveRDS(class.mig_withyear, "../mods/class.mig_withyear.rds")

```


Class, but multivariate
```{r multivariate-class-withyear}

class_withyear_prior_multivar <- prior(normal(0,5), class = 'b', 
                              coef = logMass_kg, resp = "dispersalkm") +
  prior(student_t(4, 0, 4), class = 'Intercept', resp = "dispersalkm") +
  prior(student_t(6, 0, 2.9), class = 'sd', resp = "dispersalkm") +
  prior(student_t(4, 0, 2.9), class = 'sigma', resp = "dispersalkm")+
  prior(normal(0,5), class = 'b', coef = logMass_kg, resp = "hrradius") +
  prior(student_t(4, 0, 4), class = 'Intercept', resp = "hrradius") +
  prior(student_t(6, 0, 2.9), class = 'sd', resp = "hrradius") +
  prior(student_t(4, 0, 2.9), class = 'sigma', resp = "hrradius") +
  prior(normal(0,5), class = 'b', coef = logMass_kg, resp = "Migrationkm") +
  prior(student_t(4, 0, 4), class = 'Intercept', resp = "Migrationkm") +
  prior(student_t(6, 0, 2.9), class = 'sd', resp = "Migrationkm") +
  prior(student_t(4, 0, 2.9), class = 'sigma',resp = "Migrationkm")

class_multivar <- brm(bf_dispersal_class_withyear + bf_home_range_class_withyear +
    bf_migration_class_withyear, data = traits_phylo, 
    control = list(adapt_delta = 0.995, max_treedepth = 20), 
    iter = 4000, cores = 16, prior = class_withyear_prior_multivar)

saveRDS(class_multivar, "../mods/class_multivar_withyear.rds")

```


For supplemental Figure S1: univariate models by order. Separate out orders, aves first.

```{r Order-aves-withyear}
aves <- traits_phylo %>% filter(class == "Aves")

bf_dispersal_aves_withyear <- bf(dispersal_km ~ log(Mass_kg) + 
                      (log(Mass_kg)|p|order) + dispersal_year) + lognormal()
bf_home_range_aves_withyear <- bf(hr.radius ~ log(Mass_kg)+ 
                      (log(Mass_kg)|p|order) + foraging_year) +  lognormal()
bf_migration_aves_withyear <- bf(Migration_km ~ log(Mass_kg) + 
                      (log(Mass_kg)|p|order) + migration_year) + hurdle_lognormal()

order_withyear_prior <- prior(normal(0,5),class = 'b', coef = logMass_kg) +
  prior(student_t(4, 0, 4), class = 'Intercept') +
  prior(student_t(6, 0, 2.9), class = 'sd') +
  prior(student_t(4, 0, 2.9), class = 'sigma')

aves.disp_withyear <- brm(bf_dispersal_aves_withyear, data = aves, 
                         control = list(adapt_delta = 0.995, max_treedepth=20), 
                         iter = 7000, cores = 4, prior = order_withyear_prior)

aves.for_withyear <- brm(bf_home_range_aves_withyear, data = aves, 
                        control = list(adapt_delta = 0.995, max_treedepth=20), 
                        iter = 3000, cores = 4, prior = order_withyear_prior)

aves.mig_withyear <- brm(bf_migration_aves_withyear, data = aves, 
                        control = list(adapt_delta = 0.995, max_treedepth=20), 
                        iter = 3000, cores = 4, prior = order_withyear_prior)


# saveRDS(aves.disp_withyear, "../mods/aves.disp_withyear.rds")
# saveRDS(aves.for_withyear, "../mods/aves.for_withyear.rds")
# saveRDS(aves.mig_withyear, "../mods/aves.mig_withyear.rds") 

```

mammals next
```{r Order-mammals-withyear}
mam <- traits_phylo %>% filter(class == "Mammalia")

bf_dispersal_mam_withyear <- bf(dispersal_km ~ log(Mass_kg) + 
                      (log(Mass_kg)|p|order) + dispersal_year) + lognormal()
bf_home_range_mam_withyear <- bf(hr.radius ~ log(Mass_kg)+ 
                      (log(Mass_kg)|p|order) + foraging_year) +  lognormal()
bf_migration_mam_withyear <- bf(Migration_km ~ log(Mass_kg) + 
                      (log(Mass_kg)|p|order) + migration_year) + hurdle_lognormal()

order_withyear_prior <- prior(normal(0,5),class = 'b', coef = logMass_kg) +
  prior(student_t(4, 0, 4), class = 'Intercept') +
  prior(student_t(6, 0, 2.9), class = 'sd') +
  prior(student_t(4, 0, 2.9), class = 'sigma')

mam.disp_withyear <- brm(bf_dispersal_mam_withyear, data = mam, 
                         control = list(adapt_delta = 0.995, max_treedepth=20), 
                         iter = 3000, cores = 4, prior = order_withyear_prior)

mam.for_withyear <- brm(bf_home_range_mam_withyear, data = mam, 
                        control = list(adapt_delta = 0.995, max_treedepth=20), 
                        iter = 3000, cores = 4, prior = order_withyear_prior)

# order_withyear_prior_mam <- prior(normal(0, 3), class = 'b', coef = logMass_kg) +
#   prior(student_t(3, 0, 3), class = 'Intercept') +
#   prior(student_t(3, 0, 3), class = 'sd') +
#   prior(student_t(3, 0, 3), class = 'sigma')

mam.mig_withyear <- brm(bf_migration_mam_withyear, data = mam, 
                        control = list(adapt_delta = 0.995, max_treedepth=20), 
                        iter = 4000, cores = 4, prior = order_withyear_prior)

mam.mig_withyear.2 <- brm(bf_migration_mam_withyear, data = mam, 
                        control = list(adapt_delta = 0.995, max_treedepth=20), 
                        iter = 4000, cores = 4, prior = order_withyear_prior)


# saveRDS(mam.disp_withyear, "../mods/mam.disp_withyear.rds")
# saveRDS(mam.for_withyear, "../mods/mam.for_withyear.rds")
# saveRDS(mam.mig_withyear, "../mods/mam.mig_withyear.rds") 

```




Looking at breakdown of migrators and non-migrators, starting with migrators, univariate model.
```{r migrators only, null univar mod with year}
traits_migrators <- traits_phylo %>% filter(Migration_km > 0)

bf_dispersal_null_withyear_migrators <- bf(dispersal_km ~ log(Mass_kg) + dispersal_year) + lognormal(link = "identity")
bf_home_range_null_withyear_migrators <- bf(hr.radius ~ log(Mass_kg) + foraging_year) + lognormal(link = "identity")
bf_migration_null_withyear_migrators <- bf(Migration_km ~ log(Mass_kg) + migration_year) + hurdle_lognormal(link = 'identity')


null.disp_withyear_migrators <- brm(bf_dispersal_null_withyear_migrators, data = traits_migrators, iter = 3000)
null.for_withyear_migrators <- brm(bf_home_range_null_withyear_migrators, data = traits_migrators, iter = 3000)
null.mig_withyear_migrators <- brm(bf_migration_null_withyear_migrators, data = traits_migrators, iter = 3000)

saveRDS(null.disp_withyear_migrators, "../mods/null.disp_withyear_migrators.rds")
saveRDS(null.for_withyear_migrators, "../mods/null.for_withyear_migrators.rds")
saveRDS(null.mig_withyear_migrators, "../mods/null.mig_withyear_migrators.rds") 

```

Group by class, migrators only

```{r class migrators, univariate with year}

bf_dispersal_class_withyear_migrators <- bf(dispersal_km ~ log(Mass_kg) + 
                      (log(Mass_kg)|p|class) + dispersal_year) + lognormal()
bf_home_range_class_withyear_migrators <- bf(hr.radius ~ log(Mass_kg)+ 
                      (log(Mass_kg)|p|class) + foraging_year) +  lognormal()
bf_migration_class_withyear_migrators <- bf(Migration_km ~ log(Mass_kg) + 
                      (log(Mass_kg)|p|class) + migration_year) + hurdle_lognormal()

class_withyear_prior_migrators <- prior(normal(0,5),class = 'b', coef = logMass_kg) +
  prior(student_t(4, 0, 4), class = 'Intercept') +
  prior(student_t(6, 0, 2.9), class = 'sd') +
  prior(student_t(4, 0, 2.9), class = 'sigma')

class.disp_withyear_migrators <- brm(bf_dispersal_class_withyear_migrators, data = traits_migrators, 
                         control = list(adapt_delta = 0.995, max_treedepth=20), 
                         iter = 6000, cores = 4, prior = class_withyear_prior_migrators)


class.for_withyear_migrators <- brm(bf_home_range_class_withyear_migrators, data = traits_migrators, 
                        control = list(adapt_delta = 0.995, max_treedepth=20), 
                        iter = 3000, cores = 4, prior = class_withyear_prior_migrators)



class.mig_withyear_migrators <- brm(bf_migration_class_withyear_migrators, data = traits_migrators, 
                        control = list(adapt_delta = 0.995, max_treedepth=20), 
                        iter = 3000, cores = 4, prior = class_withyear_prior_migrators)

```
```{r save migrator-only class mod}
saveRDS(class.for_withyear_migrators, "../mods/class.for_withyear_migrators.rds")

saveRDS(class.disp_withyear_migrators, "../mods/class.disp_withyear_migrators.rds")


saveRDS(class.mig_withyear_migrators, "../mods/class.mig_withyear_migrators.rds")

```

Group by trophic guild, migrators only

```{r trophic guild migrators, univariate with year}
bf_dispersal_trophic_withyear_migrators <- bf(dispersal_km ~ log(Mass_kg) + (log(Mass_kg)|diet_broadest_cat) + dispersal_year) + lognormal(link = "identity")
bf_home_range_trophic_withyear_migrators <- bf(hr.radius ~ log(Mass_kg)+ (log(Mass_kg)|diet_broadest_cat) + foraging_year) + lognormal(link = "identity")
bf_migration_trophic_withyear_migrators <- bf(Migration_km ~ log(Mass_kg) + (log(Mass_kg)|diet_broadest_cat) + migration_year) + hurdle_lognormal(link = 'identity')


trophic_year_prior_disp_migrators <- prior(normal(0,5),class = 'b', coef = logMass_kg) +
  prior(normal(0,5),class = 'b', coef = dispersal_year) +
  prior(student_t(5, 0, 3), class = 'Intercept') +
  prior(student_t(6, 0, 2.9), class = 'sd') +
  prior(student_t(5, 0, 2.9), class = 'sigma')


trophic.disp_withyear_migrators <- brm(bf_dispersal_trophic_withyear_migrators, data = traits_migrators, 
                          control = list(adapt_delta = 0.995, max_treedepth=20), 
                          iter = 3000, cores = 16,
                          prior = trophic_year_prior_disp_migrators)

trophic_year_prior_for_migrators <- prior(normal(0,5),class = 'b', coef = logMass_kg) +
  prior(normal(0,5),class = 'b', coef = foraging_year) +
  prior(student_t(5, 0, 3), class = 'Intercept') +
  prior(student_t(6, 0, 2.9), class = 'sd') +
  prior(student_t(5, 0, 2.9), class = 'sigma')


trophic.for_withyear_migrators <- brm(bf_home_range_trophic_withyear_migrators, data = traits_migrators, 
                         control = list(adapt_delta = 0.995, max_treedepth=20), iter = 4000, cores = 16,
                         prior = trophic_year_prior_for_migrators)


## works
trophic_year_prior_mig_migrators <- prior(normal(0,3),class = 'b', coef = logMass_kg) +
  prior(normal(0,3),class = 'b', coef = migration_year) +
  prior(student_t(5, 0, 3), class = 'Intercept') +
  prior(student_t(4, 0, 2.9), class = 'sd') +
  prior(student_t(4, 0, 2.9), class = 'sigma') 


trophic.mig_withyear_migrators <- brm(bf_migration_trophic_withyear_migrators, data = traits_migrators, 
                         control = list(adapt_delta = 0.995, max_treedepth=20), iter = 3000, cores = 16,
                         prior = trophic_year_prior_mig_migrators)


```

```{r save trophic migrator only mod}
saveRDS(trophic.disp_withyea_migratorsr, "../mods/trophic.disp_withyear_migrators.rds")

saveRDS(trophic.for_withyear_migrators, "../mods/trophic.for_withyear_migrators.rds")


saveRDS(trophic.mig_withyear_migrators, "../mods/trophic.mig_withyear_migrators.rds")
```


Moving on to non-migrators, univariate model, grouped by class
```{r nonmigrators only by class}
traits_nonmigrators <- traits_phylo %>% filter(Migration_km == 0)

bf_dispersal_class_withyear_nonmigrators <- bf(dispersal_km ~ log(Mass_kg) + 
                      (log(Mass_kg)|p|class) + dispersal_year) + lognormal()
bf_home_range_class_withyear_nonmigrators <- bf(hr.radius ~ log(Mass_kg)+ 
                      (log(Mass_kg)|p|class) + foraging_year) +  lognormal()


class_withyear_prior_nonmigrators <- prior(normal(0,5),class = 'b', coef = logMass_kg) +
  prior(student_t(4, 0, 4), class = 'Intercept') +
  prior(student_t(6, 0, 2.9), class = 'sd') +
  prior(student_t(4, 0, 2.9), class = 'sigma')

class.disp_withyear_nonmigrators <- brm(bf_dispersal_class_withyear_nonmigrators, data = traits_nonmigrators, 
                         control = list(adapt_delta = 0.995, max_treedepth=20), 
                         iter = 4000, cores = 4, prior = class_withyear_prior_nonmigrators)


class.for_withyear_nonmigrators <- brm(bf_home_range_class_withyear_nonmigrators, data = traits_nonmigrators, 
                        control = list(adapt_delta = 0.995, max_treedepth=20), 
                        iter = 3000, cores = 4, prior = class_withyear_prior_nonmigrators)

```
```{r save nonmigrator class mods}
saveRDS(class.disp_withyear_nonmigrators, "../mods/class.disp_withyear_nonmigrators.rds")

saveRDS(class.for_withyear_nonmigrators, "../mods/class.for_withyear_nonmigrators.rds")

```

```{r nonmigrators only by trophic guild}
bf_dispersal_trophic_withyear_nonmigrators <- bf(dispersal_km ~ log(Mass_kg) + (log(Mass_kg)|diet_broadest_cat) + dispersal_year) + lognormal(link = "identity")
bf_home_range_trophic_withyear_nonmigrators <- bf(hr.radius ~ log(Mass_kg)+ (log(Mass_kg)|diet_broadest_cat) + foraging_year) + lognormal(link = "identity")



trophic_year_prior_disp_nonmigrators <- prior(normal(0,5),class = 'b', coef = logMass_kg) +
  prior(normal(0,5),class = 'b', coef = dispersal_year) +
  prior(student_t(5, 0, 3), class = 'Intercept') +
  prior(student_t(6, 0, 2.9), class = 'sd') +
  prior(student_t(5, 0, 2.9), class = 'sigma')


trophic.disp_withyear_nonmigrators <- brm(bf_dispersal_trophic_withyear_nonmigrators, data = traits_nonmigrators, 
                          control = list(adapt_delta = 0.995, max_treedepth=20), 
                          iter = 3000, cores = 16,
                          prior = trophic_year_prior_disp_nonmigrators)

trophic_year_prior_for_nonmigrators <- prior(normal(0,5),class = 'b', coef = logMass_kg) +
  prior(normal(0,5),class = 'b', coef = foraging_year) +
  prior(student_t(5, 0, 3), class = 'Intercept') +
  prior(student_t(6, 0, 2.9), class = 'sd') +
  prior(student_t(5, 0, 2.9), class = 'sigma')

trophic.for_withyear_nonmigrators <- brm(bf_home_range_trophic_withyear_nonmigrators, data = traits_nonmigrators, 
                         control = list(adapt_delta = 0.995, max_treedepth=20), iter = 4000, cores = 16,
                         prior = trophic_year_prior_for_nonmigrators)

```
```{r save trophic nonmigrator mods}
saveRDS(trophic.disp_withyear_nonmigrators, "../mods/trophic.disp_withyear_nonmigrators.rds")

saveRDS(trophic.for_withyear_nonmigrators, "../mods/trophic.for_withyear_nonmigrators.rds")
```


```{r univariate phylo mods}

bf_dispersal_phylo_withyear <- bf(dispersal_km ~ log(Mass_kg) + 
                      (1|gr(scientific_name.x, cov = A)) + dispersal_year) + lognormal()
bf_home_range_phylo_withyear <- bf(hr.radius ~ log(Mass_kg)+ 
                      (1|gr(scientific_name.x, cov = A))  + foraging_year) +  lognormal()
bf_migration_phylo_withyear <- bf(Migration_km ~ log(Mass_kg) + 
                      (1|gr(scientific_name.x, cov = A))  + migration_year) + hurdle_lognormal()

phylo_withyear_prior_migrators <- prior(normal(0,5),class = 'b', coef = logMass_kg) +
  prior(student_t(4, 0, 4), class = 'Intercept') +
  prior(student_t(6, 0, 2.9), class = 'sd') +
  prior(student_t(4, 0, 2.9), class = 'sigma')

phylo_disp_withyear_migrators <- brm(bf_dispersal_phylo_withyear, data = traits_phylo, 
                         control = list(adapt_delta = 0.995, max_treedepth=20), data2 = list(A=A),
                         iter = 200, cores = 4, prior = phylo_withyear_prior_migrators)


phylo_for_withyear_migrators <- brm(bf_home_range_phylo_withyear, data = traits_phylo, 
                        control = list(adapt_delta = 0.995, max_treedepth=20), 
                        iter = 3000, cores = 4, prior = phylo_withyear_prior_migrators)


phylo_mig_withyear_migrators <- brm(bf_migration_class_withyear_migrators, data = traits_migrators, 
                        control = list(adapt_delta = 0.995, max_treedepth=20), 
                        iter = 3000, cores = 4, prior = phylo_withyear_prior_migrators)

phylo_for_withyear <- readRDS("../mods/phylo_for_withyear_migrators.rds")
phylo_disp_withyear <- readRDS("../mods/phylo_disp_withyear_migrators.rds")
phylo_mig_withyear <- readRDS("../mods/phylo_mig_withyear_migrators.rds")

loo(phylo_for_withyear)
bayes_R2(phylo_for_withyear)

loo(phylo_disp_withyear)
bayes_R2(phylo_disp_withyear)

loo(phylo_mig_withyear)
bayes_R2(phylo_mig_withyear)

class_for_withyear <- readRDS("../mods/class.for_withyear.rds")
class_disp_withyear <- readRDS("../mods/class.disp_withyear.rds")
class_mig_withyear <- readRDS("../mods/class.mig_withyear.rds")

loo(phylo_for_withyear,class_for_withyear)
# Model comparisons:
#                    elpd_diff se_diff
# phylo_for_withyear   0.0       0.0    55.1		
# class_for_withyear -15.4       5.6    85.9		

loo(phylo_disp_withyear,class_disp_withyear)
# Model comparisons:
#                     elpd_diff se_diff looic
# phylo_disp_withyear   0.0       0.0  2554.2
# class_disp_withyear -20.0       6.2  2594.2


loo(phylo_mig_withyear,class_mig_withyear)

# Model comparisons:
#                     elpd_diff se_diff   looic
# phylo_disp_withyear   0.0       0.0     2966.2
# class_disp_withyear -20.0       6.2     3016.3

```
