==> scripts/02_Figures.Rmd <==
---
title: "02_Figures"
author: "Sam Straus and Melissa Guzman"
date: "2023-05-01"
output: html_document
---

Read in data file that was created in the `00_Preprocessing.Rmd` script

```{r load packages and data, include=FALSE}

library(tidyverse)
library(cowplot)
library(ggplot2)
library(scales)
library(gridExtra)
library(ggpubr)
library(dplyr)
library(tidyr)
library(PNWColors)
library(gtable)
library(grid)
library(gridExtra)

# for plotting figure 1
require(holodeck)
require(scatterplot3d)
library(viridis) # color palettes
library(RColorBrewer) # color palettes

# for plotting fig 4
library(vegan) ## for PCA
# library(devtools) # to install ggbiplot from github
# install_github("vqv/ggbiplot")
library(ggbiplot) ## for PCA

# for plotting fig 5
library(tidybayes)
library(modelr)
library(brms)
library(stringr)

traits <- read.csv("../data/Traits_final_cleaned.csv", fileEncoding="latin1")

# colors used throughout
class_cols <- c(Amphibia = "#00496f", Aves = "#0f85a0", Chondrichthyes = "#edd746", 
                Mammalia = "#ed8b00", Reptilia = "#dd4124")

diet_cols <- c(Carnivore = "#4a3a3b", Herbivore = "#984136", 
                Invertivore = "#c26a7a", Omnivore = "#ecc0a1")

```

The code below is used to create the dummy hypothesis figures presented in Figure 1.
```{r Figure 1}

brewer.pal(3, "Dark")
setseed(100)

df1 <-
  sim_covar(n_obs = 500, n_vars = 3, cov = 0.6, var = 1.2, name = "movement")
scatterplot3d(df1$movement_1,df1$movement_2,df1$movement_3)

df2 <-
  df1 %>%
  sim_cat(n_groups = 3, name = "factor") %>%
  group_by(factor) %>%
  sim_discr(n_vars = 3, var = 0.9, cov = 0.3, group_means = rnorm(n=3, mean = c(1,3,5)),
            name = "trophic")


df2$factor <- as.factor(df2$factor)
colours <- RColorBrewer::brewer.pal(3, "Dark2")
df2$class_col <- colours[as.numeric(df2$factor)]


jpeg(filename = "figures/mock_3dscatter1.jpeg", width = 6, heigh = 4.5, units = "in", res = 1000)
scatterplot3d(df2$movement_1,df2$movement_2,df2$movement_3, color = df2$class_col, pch = 16,grid = FALSE,
              xlab = "Foraging scale",
              ylab = "Dispersal scale",
              zlab = "Migration scale")
par(mar = c(1,1,1,1))
dev.off()


setseed(150)
df3 <-
  sim_covar(n_obs = 500, n_vars = 3, cov = 0.0, var = 1.0, name = "movement")
scatterplot3d(df3$movement_1,df3$movement_2,df3$movement_3)

df4 <-
  df3 %>%
  sim_cat(n_groups = 3, name = "factor") %>%
  group_by(factor) %>%
  sim_discr(n_vars = 3, var = 1.3, cov = 0.0, group_means = rnorm(n=3, mean = c(3,3,3)),
            name = "trophic")

df4$factor <- as.factor(df4$factor)
colours <- RColorBrewer::brewer.pal(3, "Dark2")
df4$class_col <- colours[as.numeric(df4$factor)]

jpeg(filename = "figures/mock_3dscatter2.jpeg", width = 6, heigh = 4.5, units = "in", res = 300)
scatterplot3d(df4$movement_1,df4$movement_2,df4$movement_3, color = df4$class_col, pch = 16,grid = FALSE,
              xlab = "Foraging scale",
              ylab = "Dispersal scale",
              zlab = "Migration scale")
par(mar = c(1,1,1,1))
dev.off()


setseed(100)
df6 <- sim_covar(n_obs = 167, n_vars = 3, cov = -0.3, var = 0.8, name = "movement")
scatterplot3d(df6$movement_1,df6$movement_2,df6$movement_3)
df6 <- df6 %>% 
  mutate(movement_1 = movement_1) %>% 
  mutate(movement_2 = movement_2) %>% 
  mutate(movement_3 = movement_3-4)
df6$group <- "a"
df6$colour <- "#1B9E77"

df7 <-
  sim_covar(n_obs = 167, n_vars = 3, cov = 0.3, var = 0.8, name = "movement")
scatterplot3d(df7$movement_1,df7$movement_2,df7$movement_3)
df7$group <- "b"
df7$colour <- "#D95F02"

df8 <-
  sim_covar(n_obs = 167, n_vars = 3, cov = 0.4, var = 0.8, name = "movement")
scatterplot3d(df8$movement_1,df8$movement_2,df8$movement_3)
df8$group <- "c"
df8$colour <- "#7570B3"

df9 <- rbind(df6, df7, df8)

jpeg(filename = "figures/mock_3dscatter3.jpeg", width = 6, heigh = 4.5, units = "in", res = 300)
scatterplot3d(df9$movement_1,df9$movement_2,df9$movement_3, color = df9$colour, pch = 16, grid = FALSE,
              xlab = "Foraging scale",
              ylab = "Dispersal scale",
              zlab = "Migration scale")
par(mar = c(1,1,1,1))
dev.off()
```



The code below is used to create Figure 2, which shows the distribution of the data for each movement type and hypothesis.
```{r Figure 2 - data distributions}

# Figure 2 top row: group by movement medium
# this is only used once, so assigning colors here rather than in chunk 1
media_cols <- c(air = "#fbcc74", land = "#ba783e", water = "#4f412b")


# hist by media
traits$media_foraging[which(traits$media_foraging =="marine")] <- "water"
traits$media_foraging[which(traits$media_foraging =="aquatic")] <- "water"

media_for_distr <- traits %>%
  filter(media_foraging == "air" | media_foraging == "land" |
           media_foraging == "water") %>%
  ggplot(aes(x = media_foraging, y = hr.radius))+
    geom_violin(aes(color = media_foraging, fill = media_foraging))  + 
  geom_boxplot(width = 0.2)+
  scale_color_manual(values = media_cols, name = "Medium")+
  scale_fill_manual(values = media_cols, name = "Medium") +
  scale_y_log10(labels = trans_format("log10", math_format(10^.x)))+
  theme(legend.position = "none")+
  cowplot::theme_cowplot()+
  ylab("Foraging radius (km)")+
  # xlab("Foraging medium")+
  theme(axis.title.x=element_blank(), axis.text.x = element_text(angle = 45, vjust = 1, hjust=1))+
  theme(legend.position = "none")

traits$media_dispersal[which(traits$media_dispersal =="marine")] <- "water"
traits$media_dispersal[which(traits$media_dispersal =="aquatic")] <- "water"

media_disp_distr <- traits %>%
  filter(media_dispersal == "air" | media_dispersal == "land" |
           media_dispersal == "water") %>%
  ggplot(aes(x = media_dispersal, y = dispersal_km))+
    geom_violin(aes(color = media_dispersal, fill = media_dispersal))  + 
  geom_boxplot(width = 0.2)+
  scale_color_manual(values = media_cols, name = "Medium")+
  scale_fill_manual(values = media_cols, name = "Medium") +
  scale_y_log10(labels = trans_format("log10", math_format(10^.x)))+
  theme(legend.position = "none")+
  cowplot::theme_cowplot()+
  ylab("Dispersal distance (km)")+
  # xlab("Dispersal medium")+
  theme(axis.title.x=element_blank(), axis.text.x = element_text(angle = 45, vjust = 1, hjust=1))+
  theme(legend.position = "none")


traits$media_migration[which(traits$media_migration =="marine")] <- "water"
traits$media_migration[which(traits$media_migration =="aquatic")] <- "water"

media_mig_distr <- traits %>%
  filter(media_migration == "air" | media_migration == "land" |
           media_migration == "water") %>%
ggplot(aes(x = media_migration, y = (Migration_km + 1)))+
    geom_violin(aes(color = media_migration, fill = media_migration))  + 
  geom_boxplot(width = 0.2)+
  scale_color_manual(values = media_cols, name = "Medium")+
  scale_fill_manual(values = media_cols, name = "Medium") +
  scale_y_log10(labels = trans_format("log10", math_format(10^.x)))+
  theme(legend.position = "none")+
  cowplot::theme_cowplot()+
  ylab("Migration distance (km + 1)")+
  # xlab("Migration medium")+
  theme(axis.title.x=element_blank(), axis.text.x = element_text(angle = 45, vjust = 1, hjust=1))+
  theme(legend.position = "none")


# Figure 2 mid row: group by taxonomic class


class_for_distr <- ggplot(traits, aes(x = class, y = hr.radius)) + 
  geom_violin(aes(color = class, fill = class))  + 
  geom_boxplot(width = 0.2)+
  scale_color_manual(values = class_cols, name = "Class")+
  scale_fill_manual(values = class_cols, name = "Class") +
  scale_y_log10(labels = trans_format("log10", math_format(10^.x)))+
  theme(legend.position = "none")+
  cowplot::theme_cowplot()+
  ylab("Foraging radius (km)")+
  # xlab("Class")+
  theme(axis.title.x=element_blank(), axis.text.x = element_text(angle = 45, vjust = 1, hjust=1))+
  theme(legend.position = "none")


class_disp_distr <- ggplot(traits, aes(x = class, y = dispersal_km)) + 
  geom_violin(aes(color = class, fill = class))  + 
  geom_boxplot(width = 0.2)+
  scale_color_manual(values = class_cols, name = "Class")+
  scale_fill_manual(values = class_cols, name = "Class") +
  scale_y_log10(labels = trans_format("log10", math_format(10^.x)))+
  theme(legend.position = "none")+
  cowplot::theme_cowplot()+
  ylab("Dispersal distance (km)")+
  # xlab("Class")+
  theme(axis.title.x=element_blank(), axis.text.x = element_text(angle = 45, vjust = 1, hjust=1))+
  theme(legend.position = "none")

class_mig_distr <- ggplot(traits, aes(x = class, y = (Migration_km + 1))) + 
  geom_violin(aes(color = class, fill = class))  + 
  geom_boxplot(width = 0.2)+
  scale_color_manual(values = class_cols, name = "Class")+
  scale_fill_manual(values = class_cols, name = "Class") +
  scale_y_log10(labels = trans_format("log10", math_format(10^.x)))+
  theme(legend.position = "none")+
  cowplot::theme_cowplot()+
  ylab("Migration distance (km + 1)")+
  # xlab("Class")+
  theme(axis.title.x=element_blank(), axis.text.x = element_text(angle = 45, vjust = 1, hjust=1))+
  theme(legend.position = "none")

distr_by_class <- ggarrange(class_for_distr, class_disp_distr, class_mig_distr, nrow = 1, 
                           labels = c("a", "b", "c"))


# Figure 2 bottom panel, group by trophic guild


trophic_for_distr <-  traits %>% 
  filter(diet_broadest_cat != "") %>%
  ggplot(aes(x = diet_broadest_cat, y = hr.radius)) +
  geom_violin(aes(color = diet_broadest_cat, fill = diet_broadest_cat))  + 
  geom_boxplot(width = 0.2)+
  scale_color_manual(values = diet_cols, name = "Trophic Guild")+
  scale_fill_manual(values = diet_cols, name = "Trophic Guild") +
  scale_y_log10(labels = trans_format("log10", math_format(10^.x)))+
  theme(legend.position = "none")+
  cowplot::theme_cowplot()+
  ylab("Foraging radius (km)")+
  # xlab("Trophic Guild")+
  theme(axis.title.x=element_blank(), axis.text.x = element_text(angle = 45, vjust = 1, hjust=1))+
  theme(legend.position = "none")

trophic_disp_distr <-  traits %>% 
  filter(diet_broadest_cat != "") %>%
  ggplot(aes(x = diet_broadest_cat, y = dispersal_km)) +
  geom_violin(aes(color = diet_broadest_cat, fill = diet_broadest_cat))  + 
  geom_boxplot(width = 0.2)+
  scale_color_manual(values = diet_cols, name = "Trophic Guild")+
  scale_fill_manual(values = diet_cols, name = "Trophic Guild") +
  scale_y_log10(labels = trans_format("log10", math_format(10^.x)))+
  theme(legend.position = "none")+
  cowplot::theme_cowplot()+
  ylab("Dispersal distance (km)")+
  # xlab("Trophic Guild")+
  theme(axis.title.x=element_blank(), axis.text.x = element_text(angle = 45, vjust = 1, hjust=1))+
  theme(legend.position = "none")


trophic_mig_distr <-  traits %>% 
  filter(diet_broadest_cat != "") %>%
  ggplot(aes(x = diet_broadest_cat, y = (Migration_km + 1))) +
  geom_violin(aes(color = diet_broadest_cat, fill = diet_broadest_cat))  + 
  geom_boxplot(width = 0.2)+
  scale_color_manual(values = diet_cols, name = "Trophic Guild")+
  scale_fill_manual(values = diet_cols, name = "Trophic Guild") +
  scale_y_log10(labels = trans_format("log10", math_format(10^.x)))+
  theme(legend.position = "none")+
  cowplot::theme_cowplot()+
  ylab("Migration distance (km + 1)")+
  # xlab("Trophic Guild")+
  theme(axis.title.x=element_blank(), axis.text.x = element_text(angle = 45, vjust = 1, hjust=1))+
  theme(legend.position = "none")

##  make the widths all the same
# Get the gtables
gA <- ggplotGrob(class_for_distr)
gB <- ggplotGrob(class_disp_distr)
gC <- ggplotGrob(class_mig_distr)
gD <- ggplotGrob(trophic_for_distr)
gE <- ggplotGrob(trophic_disp_distr)
gF <- ggplotGrob(trophic_mig_distr)

gG <- ggplotGrob(media_for_distr)
gH <- ggplotGrob(media_disp_distr)
gI <- ggplotGrob(media_mig_distr)

gA$heights <- gC$heights
gB$heights <- gC$heights
gD$heights <- gC$heights
gE$heights <- gC$heights
gF$heights <- gC$heights
gG$heights <- gC$heights
gH$heights <- gC$heights
gI$heights <- gC$heights



while (!is.null(dev.list()))  dev.off()
jpeg(filename = "../figures/Figure02_07Aug2023.jpeg", width = 12, height = 11, units = 'in', res = 1000)
grid.arrange(gG, gH, gI, gA, gB, gC, gD,gE, gF, nrow = 3, padding = 2)
dev.off()

```

Create PCA plots for class and trophic guild. We plotted the standardized mass and movement values for three movement types, and also the residual of movement distance ~ body mass for each type. This chode chunk also pulls PCA loadings, presented in the results section and Table S2.

```{r PCA plots}

traits$log.mass <- log10(traits$Mass_kg)
traits$log.mig <- log10(traits$Migration_km)
traits$log.disp <- log10(traits$dispersal_km)
traits$log.for <- log10(traits$hr.radius)
 

traits <- traits[-which(traits$log.mig == -Inf),]
traits <- traits[which(!is.na(traits$log.for)),]
traits <- traits[which(!is.na(traits$log.disp)),]


names(traits)
test <- decostand(traits[, 25:27], method = "standardize")


v <- var(test)
#View(v)
diag(v) #check variance of traits in same order of magnitude (else need to transform)


# z <- prcomp(test)  
z <- prcomp(test)
loadings.raw <- z$rotation


pca_class <- ggbiplot(z, ellipse = TRUE, choices = c(1,2), groups = traits$class) +
  scale_color_manual(values = class_cols) + 
  xlab("PC1 (61.9%)")+
  ylab("PC2 (22.9%)")+
  theme_cowplot()

pca_diet <- ggbiplot(z, ellipse = TRUE, choices = c(1,2), groups = traits$diet_broadest_cat) + 
  scale_color_manual(values = diet_cols)+
  xlab("PC1 (61.9%)")+
  ylab("PC2 (22.9%)")+
  theme_cowplot()


pca1 <- ggarrange(pca_class, pca_diet, nrow = 2, legend = "none", labels = c("a", "c"))


## residuals
disp.bs <- lm(log.disp ~ log.mass, data = traits)
#plot(log.disp ~ log.mass, data = traits)
traits$disp.resid <- resid(disp.bs)

mig.bs <- lm(log.mig ~ log.mass, data = traits)
#plot(log.mig ~ log.mass, data = traits)
traits$mig.resid <- resid(mig.bs)

for.bs <- lm(log.for ~ log.mass, data = traits)
#plot(log.for ~ log.mass, data = traits)
traits$for.resid <- resid(for.bs)

#names(traits)
test2 <- decostand(traits[, 28:30], method = "standardize")

v2 <- var(test2)
#View(v2)

z2 <- prcomp(test2)
loadings.z2 <- z2$rotation

ggbiplot(z2, choices = c(1,2), cex = 0.7)

pca_class_res <- ggbiplot(z2, ellipse = TRUE, choices = c(1,2), groups = traits$class) +
  scale_color_manual(values = class_cols) + 
  xlab("PC1 (61.4%)")+
  ylab("PC2 (20.5%)")+
  theme_cowplot()

pca_diet_res <- ggbiplot(z2, ellipse = TRUE, choices = c(1,2), groups = traits$diet_broadest_cat) + 
  scale_color_manual(values = diet_cols)+
  xlab("PC1 (61.4%)")+
  ylab("PC2 (20.5%)")+
  theme_cowplot()

# Get the gtables

# Get the gtables
gA <- ggplotGrob(pca_class)
gB <- ggplotGrob(pca_diet)
gC <- ggplotGrob(pca_class_res)
gD <- ggplotGrob(pca_diet_res)

gB$widths <- gA$widths
gC$widths <- gA$widths
gD$widths <- gA$widths

gA$heights <- gC$heights
gB$heights <- gC$heights
gD$heights <- gC$heights

# Arrange the two charts.
# The legend boxes are centered

while (!is.null(dev.list()))  dev.off()
jpeg(file = "../figures/Figure3_07Aug2023.jpeg", width = 10, height = 10, units = "in", res = 1000)
grid.arrange(gA, gC, gB, gD, nrow = 2, padding = 2)
dev.off()


```


The following code chunk generates Figure 5.

```{r Figure 5 class}

class_dispersal <- readRDS("../mods/class.disp_withyear.rds")
class_foraging <- readRDS("../mods/class.for_withyear.rds")
class_migration <- readRDS("../mods/class.mig_withyear.rds")


pred_mod_class_dispersal <- traits %>%
  group_by(class) %>%
  data_grid(Mass_kg = seq_range(Mass_kg, n = 320), 
            dispersal_year = seq(min(dispersal_year), max(dispersal_year), by = 7)) %>% 
  add_epred_draws(class_dispersal)

pred_mod_class_foraging <- traits %>%
  group_by(class) %>%
  data_grid(Mass_kg = seq_range(Mass_kg, n = 320), 
            foraging_year = seq(min(foraging_year, na.rm = TRUE), max(foraging_year, na.rm = TRUE), by = 7)) %>% 
  add_epred_draws(class_foraging)

pred_mod_class_migration <- traits %>%
  group_by(class) %>%
  data_grid(Mass_kg = seq_range(Mass_kg, n = 320), 
            migration_year = seq(min(migration_year, na.rm = TRUE), max(migration_year, na.rm=TRUE), by = 7)) %>% 
  add_epred_draws(class_migration)



class_foraging_plot <- ggplot() +
  geom_point(data = traits, aes(x = Mass_kg, y = log10(hr.radius), color = ordered(class))) +
  stat_lineribbon(data = pred_mod_class_foraging, 
                  aes(y = log10(.epred), x = Mass_kg, color = ordered(class), fill = ordered(class)), 
                  .width = c(.95), alpha = 0.4) +
  scale_color_manual(values = class_cols, name = "Class")+
  scale_fill_manual(values = class_cols, name = "Class") + 
  scale_x_log10(breaks = scales::trans_breaks("log10", function(x) 10^x),
                labels = scales::trans_format("log10", scales::math_format(10^.x)))+
  scale_y_continuous(labels = scales::math_format())+
  # scale_y_log10(breaks = scales::trans_breaks("log10", function(x) 10^x),
  #               labels = scales::trans_format("log10", scales::math_format(10^.x)))+
  cowplot::theme_cowplot()+
  xlab("Body mass (kg)")+
  ylab("Foraging radius (km)")


class_dispersal_plot <- ggplot()+
  geom_point(data = traits, aes(x = Mass_kg, y = log10(dispersal_km), color = ordered(class))) +
  stat_lineribbon(data =pred_mod_class_dispersal,
                  aes(y = log10(.epred), x = Mass_kg,
                      color = ordered(class), fill = ordered(class)),
                  .width = 0.95, alpha = 0.4)+
  scale_color_manual(values = class_cols, name = "Class")+
  scale_fill_manual(values = class_cols, name = "Class") + 
  scale_x_log10(breaks = scales::trans_breaks("log10", function(x) 10^x),
                labels = scales::trans_format("log10", scales::math_format(10^.x)))+
  scale_y_continuous(labels = scales::math_format())+
  # scale_y_log10(breaks = scales::trans_breaks("log10", function(x) 10^x),
  #               labels = scales::trans_format("log10", scales::math_format(10^.x)))+
  cowplot::theme_cowplot()+
  xlab("Body mass (kg)")+
  ylab("Dispersal distance (km)")


class_migration_plot <- ggplot() +
  geom_point(data = traits, aes(x = Mass_kg, y = log10(Migration_km), color = ordered(class))) +
  stat_lineribbon(data = pred_mod_class_migration, 
                  aes(y = log10(.epred), x = Mass_kg, color = ordered(class),fill =ordered(class)), 
                  .width = c(.95), alpha = 0.4) +
  scale_color_manual(values = class_cols, name = "Class")+
  scale_fill_manual(values = class_cols, name = "Class") + 
  scale_x_log10(breaks = scales::trans_breaks("log10", function(x) 10^x),
                labels = scales::trans_format("log10", scales::math_format(10^.x)))+
  # scale_y_log10(breaks = scales::trans_breaks("log10", function(x) 10^x),
  #               labels = scales::trans_format("log10", scales::math_format(10^.x)))+
  scale_y_continuous(labels = scales::math_format())+
  cowplot::theme_cowplot()+
  xlab("Body mass (kg)")+
  ylab("Migration distance (km)")
```

```{r trophic guild figure 5s}
### trophic guild models
traits$diet_broadest_cat <- as.factor(traits$diet_broadest_cat)

trophic_foraging <- readRDS("../mods/trophic.for_withyear.rds")
trophic_dispersal <- readRDS("../mods/trophic.disp_withyear.rds")
trophic_migration <- readRDS("../mods/trophic.mig_withyear.rds")


pred_mod_trophic_foraging <- traits %>%
  group_by(diet_broadest_cat) %>%
  data_grid(Mass_kg = seq_range(Mass_kg, n = 320), 
            foraging_year = seq(min(foraging_year, na.rm = TRUE), max(foraging_year, na.rm = TRUE), by = 7)) %>% 
  add_epred_draws(trophic_foraging)

pred_mod_trophic_dispersal <- traits %>%
  group_by(diet_broadest_cat) %>%
  data_grid(Mass_kg = seq_range(Mass_kg, n = 320), 
            dispersal_year = seq(min(dispersal_year), max(dispersal_year), by = 7)) %>% 
  add_epred_draws(trophic_dispersal)

pred_mod_trophic_migration <- traits %>%
  group_by(diet_broadest_cat) %>%
  data_grid(Mass_kg = seq_range(Mass_kg, n = 320), 
            migration_year = seq(min(migration_year, na.rm = TRUE), max(migration_year, na.rm=TRUE), by = 7)) %>% 
  add_epred_draws(trophic_migration)



trophic_foraging_plot <- ggplot() +
  geom_point(data = traits, aes(x = Mass_kg, y = log10(hr.radius), color = ordered(diet_broadest_cat))) +
  stat_lineribbon(data = pred_mod_trophic_foraging, 
                  aes(y = log10(.epred), x = Mass_kg, color = ordered(diet_broadest_cat), fill = ordered(diet_broadest_cat)), 
                  .width = c(.95), alpha = 0.4) +
  scale_color_manual(values = diet_cols, name = "Trophic Guild")+
  scale_fill_manual(values = diet_cols, name = "Trophic Guild") +
  scale_x_log10(breaks = scales::trans_breaks("log10", function(x) 10^x),
                labels = scales::trans_format("log10", scales::math_format(10^.x)))+
  scale_y_continuous(labels = scales::math_format())+
  # scale_y_log10(breaks = scales::trans_breaks("log10", function(x) 10^x),
  #               labels = scales::trans_format("log10", scales::math_format(10^.x)))+
  cowplot::theme_cowplot()+
  xlab("Body mass (kg)")+
  ylab("Foraging radius (km)")


trophic_dispersal_plot <- ggplot()+
  geom_point(data = traits, aes(x = Mass_kg, y = log10(dispersal_km), 
                                color = ordered(diet_broadest_cat))) +
  stat_lineribbon(data = pred_mod_trophic_dispersal,
                  aes(y = log10(.epred), x = Mass_kg,
                      color = ordered(diet_broadest_cat), 
                      fill = ordered(diet_broadest_cat)),
                  .width = 0.95, alpha = 0.4)+
  scale_color_manual(values = diet_cols, name = "Trophic Guild")+
  scale_fill_manual(values = diet_cols, name = "Trophic Guild") +
  scale_x_log10(breaks = scales::trans_breaks("log10", function(x) 10^x),
                labels = scales::trans_format("log10", scales::math_format(10^.x)))+
  scale_y_continuous(labels = scales::math_format())+
  # scale_y_log10(breaks = scales::trans_breaks("log10", function(x) 10^x),
  #               labels = scales::trans_format("log10", scales::math_format(10^.x)))+
  cowplot::theme_cowplot()+
  xlab("Body mass (kg)")+
  ylab("Dispersal distance (km)")

trophic_migration_plot <- ggplot() +
  geom_point(data = traits, aes(x = Mass_kg, y = log10(Migration_km), color = ordered(diet_broadest_cat))) +
stat_lineribbon(data = pred_mod_trophic_migration[(pred_mod_trophic_migration$.epred>0),],
                  aes(y = log10(.epred), x = Mass_kg, 
                      color = ordered(diet_broadest_cat), 
                      fill = ordered(diet_broadest_cat)), 
                  .width = c(.95), alpha = 0.4) +
  scale_color_manual(values = diet_cols, name = "Trophic Guild")+
  scale_fill_manual(values = diet_cols, name = "Trophic Guild") + 
  scale_x_log10(breaks = scales::trans_breaks("log10", function(x) 10^x),
                labels = scales::trans_format("log10", scales::math_format(10^.x)))+
  # scale_y_log10(breaks = scales::trans_breaks("log10", function(x) 10^x),
  #               labels = scales::trans_format("log10", scales::math_format(10^.x)))+
  scale_y_continuous(labels = scales::math_format())+
  cowplot::theme_cowplot()+
  xlab("Body mass (kg)")+
  ylab("Migration distance (km)")

```

```{r render fig 5}
# 
# class_figs <- ggarrange(class_foraging_plot, class_dispersal_plot, class_migration_plot, nrow = 1, 
#                        common.legend = T, legend = "right")
# 
# diet_figs <- ggarrange(trophic_foraging_plot, trophic_dispersal_plot, trophic_migration_plot, nrow = 1, 
#                        common.legend = T, legend = "right")

# Get the gtables
gA <- ggplotGrob(class_foraging_plot)
gB <- ggplotGrob(class_dispersal_plot)
gC <- ggplotGrob(class_migration_plot)
gD <- ggplotGrob(trophic_foraging_plot)
gE <- ggplotGrob(trophic_dispersal_plot)
gF <- ggplotGrob(trophic_migration_plot)


gA$widths <- gC$widths
gB$widths <- gC$widths
gD$widths <- gC$widths
gE$widths <- gC$widths
gF$widths <- gC$widths
  


# Figure5 <- ggarrange(class_figs, diet_figs, nrow=2, common.legend=FALSE)

while (!is.null(dev.list()))  dev.off()
jpeg(filename = "../figures/Fig5_epreds_07Aug2023.jpeg", width = 10, height = 10, units = 'in', res = 1000)
grid.arrange(gA, gD, gB, gE,gC, gF, nrow = 3, padding = 2)
dev.off()

```

From here on out, code chunks are used to generate the figures and tables presented in the supplemental materials.

Figure S1. Top panel, birds
```{r Fig S1 birds}
aves <- traits %>% filter(class == "Aves")


aves_dispersal <- readRDS("../mods/aves.disp_withyear.rds")
aves_foraging <- readRDS("../mods/aves.for_withyear.rds")
aves_migration <- readRDS("../mods/aves.mig_withyear.rds")


pred_mod_order_dispersal_aves <- aves %>%
  group_by(order) %>%
  data_grid(Mass_kg = seq_range(Mass_kg, n = 159), 
            dispersal_year = seq(min(dispersal_year), max(dispersal_year), by = 7)) %>% 
  add_epred_draws(aves_dispersal)

pred_mod_order_foraging_aves <- aves %>%
  group_by(order) %>%
  data_grid(Mass_kg = seq_range(Mass_kg, n = 159), 
            foraging_year = seq(min(foraging_year, na.rm = TRUE), max(foraging_year, na.rm = TRUE), by = 7)) %>% 
  add_epred_draws(aves_foraging)

pred_mod_order_migration_aves <- aves %>%
  group_by(order) %>%
  data_grid(Mass_kg = seq_range(Mass_kg, n = 159), 
            migration_year = seq(min(migration_year, na.rm = TRUE), max(migration_year, na.rm=TRUE), by = 7)) %>% 
  add_epred_draws(aves_migration)

aves_foraging_plot <- ggplot() +
  geom_point(data = aves, aes(x = Mass_kg, y = log10(hr.radius), color = ordered(order))) +
  stat_lineribbon(data = pred_mod_order_foraging_aves, 
                  aes(y = log10(.epred), x = Mass_kg, color = ordered(order), fill = ordered(order)), 
                  .width = c(.95), alpha = 0.4) +
  # scale_color_manual(values = class_cols, name = "Order")+
  # scale_fill_manual(values = class_cols, name = "Order") + 
  scale_x_log10(breaks = scales::trans_breaks("log10", function(x) 10^x),
                labels = scales::trans_format("log10", scales::math_format(10^.x)))+
  scale_y_continuous(labels = scales::math_format())+
  cowplot::theme_cowplot()+
  xlab("Body mass (kg)")+
  ylab("Foraging radius (km)")+
  labs(fill = "Order", color = "Order")


aves_dispersal_plot <- ggplot()+
  geom_point(data = aves, aes(x = Mass_kg, y = log10(dispersal_km), color = ordered(order))) +
  stat_lineribbon(data =pred_mod_order_dispersal_aves,
                  aes(y = log10(.epred), x = Mass_kg,
                      color = ordered(order), fill = ordered(order)),
                  .width = 0.95, alpha = 0.4)+
  # scale_color_manual(values = class_cols, name = "Order")+
  # scale_fill_manual(values = class_cols, name = "Order") + 
  scale_x_log10(breaks = scales::trans_breaks("log10", function(x) 10^x),
                labels = scales::trans_format("log10", scales::math_format(10^.x)))+
  scale_y_continuous(labels = scales::math_format())+
  cowplot::theme_cowplot()+
  xlab("Body mass (kg)")+
  ylab("Dispersal distance (km)")+
  labs(fill = "Order", color = "Order")


aves_migration_plot <- ggplot() +
  geom_point(data = aves, aes(x = Mass_kg, y = log10(Migration_km), color = ordered(order))) +
  stat_lineribbon(data = pred_mod_order_migration_aves, 
                  aes(y = log10(.epred), x = Mass_kg, color = ordered(order),fill =ordered(order)), 
                  .width = c(.95), alpha = 0.4) +
  # scale_color_manual(values = class_cols, name = "Order")+
  # scale_fill_manual(values = class_cols, name = "Order") + 
  scale_x_log10(breaks = scales::trans_breaks("log10", function(x) 10^x),
                labels = scales::trans_format("log10", scales::math_format(10^.x)))+
  scale_y_continuous(labels = scales::math_format())+
  cowplot::theme_cowplot()+
  xlab("Body mass (kg)")+
  ylab("Migration distance (km)")+
  labs(fill = "Order", color = "Order")


gA <- ggplotGrob(aves_foraging_plot)
gB <- ggplotGrob(aves_dispersal_plot)
gC <- ggplotGrob(aves_migration_plot)


while (!is.null(dev.list()))  dev.off()
jpeg(filename = "../figures/aves_order_05May2023.jpeg", width = 16, height = 4, units = 'in', res = 1000)
grid.arrange(gA, gB, gC, nrow = 1, padding = 2)
dev.off()
```


Figure S1. Bottom panel, mammals
```{r Fig S1 mammals}
mam <- traits %>% filter(class == "Mammalia")


mam_dispersal <- readRDS("../mods/mam.disp_withyear.rds")
mam_foraging <- readRDS("../mods/mam.for_withyear.rds")
mam_migration <- readRDS("../mods/mam.mig_withyear.rds")


pred_mod_order_dispersal_mam <- mam %>%
  group_by(order) %>%
  data_grid(Mass_kg = seq_range(Mass_kg, n = 115), 
            dispersal_year = seq(min(dispersal_year), max(dispersal_year), by = 7)) %>% 
  add_epred_draws(mam_dispersal)

pred_mod_order_foraging_mam <- mam %>%
  group_by(order) %>%
  data_grid(Mass_kg = seq_range(Mass_kg, n = 115), 
            foraging_year = seq(min(foraging_year, na.rm = TRUE), max(foraging_year, na.rm = TRUE), by = 7)) %>% 
  add_epred_draws(mam_foraging)

pred_mod_order_migration_mam <- mam %>%
  group_by(order) %>%
  data_grid(Mass_kg = seq_range(Mass_kg, n = 115), 
            migration_year = seq(min(migration_year, na.rm = TRUE), max(migration_year, na.rm=TRUE), by = 7)) %>% 
  add_epred_draws(mam_migration)

mam_foraging_plot <- ggplot() +
  geom_point(data = mam, aes(x = Mass_kg, y = log10(hr.radius), color = ordered(order))) +
  stat_lineribbon(data = pred_mod_order_foraging_mam, 
                  aes(y = log10(.epred), x = Mass_kg, color = ordered(order), fill = ordered(order)), 
                  .width = c(.95), alpha = 0.4) +
  # scale_color_manual(values = class_cols, name = "Order")+
  # scale_fill_manual(values = class_cols, name = "Order") + 
  scale_x_log10(breaks = scales::trans_breaks("log10", function(x) 10^x),
                labels = scales::trans_format("log10", scales::math_format(10^.x)))+
  scale_y_continuous(labels = scales::math_format())+
  cowplot::theme_cowplot()+
  xlab("Body mass (kg)")+
  ylab("Foraging radius (km)")+
  labs(fill = "Order", color = "Order")


mam_dispersal_plot <- ggplot()+
  geom_point(data = mam, aes(x = Mass_kg, y = log10(dispersal_km), color = ordered(order))) +
  stat_lineribbon(data =pred_mod_order_dispersal_mam,
                  aes(y = log10(.epred), x = Mass_kg,
                      color = ordered(order), fill = ordered(order)),
                  .width = 0.95, alpha = 0.4)+
  # scale_color_manual(values = class_cols, name = "Order")+
  # scale_fill_manual(values = class_cols, name = "Order") + 
  scale_x_log10(breaks = scales::trans_breaks("log10", function(x) 10^x),
                labels = scales::trans_format("log10", scales::math_format(10^.x)))+
  scale_y_continuous(labels = scales::math_format())+
  cowplot::theme_cowplot()+
  xlab("Body mass (kg)")+
  ylab("Dispersal distance (km)")+
  labs(fill = "Order", color = "Order")


mam_migration_plot <- ggplot() +
  geom_point(data = mam, aes(x = Mass_kg, y = log10(Migration_km), color = ordered(order))) +
  stat_lineribbon(data = pred_mod_order_migration_mam, 
                  aes(y = log10(.epred), x = Mass_kg, color = ordered(order),fill =ordered(order)), 
                  .width = c(.95), alpha = 0.4) +
  # scale_color_manual(values = class_cols, name = "Order")+
  # scale_fill_manual(values = class_cols, name = "Order") + 
  scale_x_log10(breaks = scales::trans_breaks("log10", function(x) 10^x),
                labels = scales::trans_format("log10", scales::math_format(10^.x)))+
  scale_y_continuous(labels = scales::math_format())+
  cowplot::theme_cowplot()+
  xlab("Body mass (kg)")+
  ylab("Migration distance (km)")+
  labs(fill = "Order", color = "Order")

gA <- ggplotGrob(mam_foraging_plot)
gB <- ggplotGrob(mam_dispersal_plot)
gC <- ggplotGrob(mam_migration_plot)


while (!is.null(dev.list()))  dev.off()
jpeg(filename = "../figures/mam_order_05May2023.jpeg", width = 16, height = 4, units = 'in', res = 1000)
grid.arrange(gA, gB, gC, nrow = 1, padding = 2)
dev.off()
```


To create Figure S2, modify the code from Figure 5 above using only migratory species.

```{r Fig S2 migrators only}
traits_migrators <- traits %>% filter(Migration_km > 0)

## class
class_dispersal_migrators <- readRDS("../mods/class.disp_withyear_migrators.rds")
class_foraging_migrators <- readRDS("../mods/class.for_withyear_migrators.rds")
class_migration_migrators <- readRDS("../mods/class.mig_withyear_migrators.rds")


pred_mod_class_dispersal_migrators <- traits_migrators %>%
  group_by(class) %>%
  data_grid(Mass_kg = seq_range(Mass_kg, n = 183), 
            dispersal_year = seq(min(dispersal_year), max(dispersal_year), by = 7)) %>% 
  add_epred_draws(class_dispersal_migrators)

pred_mod_class_foraging_migrators <- traits_migrators %>%
  group_by(class) %>%
  data_grid(Mass_kg = seq_range(Mass_kg, n = 183), 
            foraging_year = seq(min(foraging_year, na.rm = TRUE), max(foraging_year, na.rm = TRUE), by = 7)) %>% 
  add_epred_draws(class_foraging_migrators)

pred_mod_class_migration_migrators <- traits_migrators %>%
  group_by(class) %>%
  data_grid(Mass_kg = seq_range(Mass_kg, n = 183), 
            migration_year = seq(min(migration_year, na.rm = TRUE), max(migration_year, na.rm=TRUE), by = 7)) %>% 
  add_epred_draws(class_migration_migrators)

class_foraging_plot_migrators <- ggplot() +
  geom_point(data = traits_migrators, aes(x = Mass_kg, y = log10(hr.radius), color = ordered(class))) +
  stat_lineribbon(data = pred_mod_class_foraging_migrators, 
                  aes(y = log10(.epred), x = Mass_kg, color = ordered(class), fill = ordered(class)), 
                  .width = c(.95), alpha = 0.4) +
  scale_color_manual(values = class_cols, name = "Class")+
  scale_fill_manual(values = class_cols, name = "Class") + 
  scale_x_log10(breaks = scales::trans_breaks("log10", function(x) 10^x),
                labels = scales::trans_format("log10", scales::math_format(10^.x)))+
  scale_y_continuous(labels = scales::math_format())+
  cowplot::theme_cowplot()+
  xlab("Body mass (kg)")+
  ylab("Foraging radius (km)")


class_dispersal_plot_migrators <- ggplot()+
  geom_point(data = traits_migrators, aes(x = Mass_kg, y = log10(dispersal_km), color = ordered(class))) +
  stat_lineribbon(data =pred_mod_class_dispersal_migrators,
                  aes(y = log10(.epred), x = Mass_kg,
                      color = ordered(class), fill = ordered(class)),
                  .width = 0.95, alpha = 0.4)+
  scale_color_manual(values = class_cols, name = "Class")+
  scale_fill_manual(values = class_cols, name = "Class") + 
  scale_x_log10(breaks = scales::trans_breaks("log10", function(x) 10^x),
                labels = scales::trans_format("log10", scales::math_format(10^.x)))+
  scale_y_continuous(labels = scales::math_format())+
  cowplot::theme_cowplot()+
  xlab("Body mass (kg)")+
  ylab("Dispersal distance (km)")


class_migration_plot_migrators <- ggplot() +
  geom_point(data = traits_migrators, aes(x = Mass_kg, y = log10(Migration_km), color = ordered(class))) +
  stat_lineribbon(data = pred_mod_class_migration_migrators, 
                  aes(y = log10(.epred), x = Mass_kg, color = ordered(class),fill =ordered(class)), 
                  .width = c(.95), alpha = 0.4) +
  scale_color_manual(values = class_cols, name = "Class")+
  scale_fill_manual(values = class_cols, name = "Class") + 
  scale_x_log10(breaks = scales::trans_breaks("log10", function(x) 10^x),
                labels = scales::trans_format("log10", scales::math_format(10^.x)))+
  scale_y_continuous(labels = scales::math_format())+
  cowplot::theme_cowplot()+
  xlab("Body mass (kg)")+
  ylab("Migration distance (km)")



## trophic guild
traits$diet_broadest_cat <- as.factor(traits$diet_broadest_cat)

trophic_foraging_migrators <- readRDS("../mods/trophic.for_withyear_migrators.rds")
trophic_dispersal_migrators <- readRDS("../mods/trophic.disp_withyear_migrators.rds")
trophic_migration_migrators <- readRDS("../mods/trophic.mig_withyear_migrators.rds")

pred_mod_trophic_foraging_migrators <- traits_migrators %>%
  group_by(diet_broadest_cat) %>%
  data_grid(Mass_kg = seq_range(Mass_kg, n = 183), 
            foraging_year = seq(min(foraging_year, na.rm = TRUE), max(foraging_year, na.rm = TRUE), by = 7)) %>% 
  add_epred_draws(trophic_foraging_migrators)

pred_mod_trophic_dispersal_migrators <- traits_migrators %>%
  group_by(diet_broadest_cat) %>%
  data_grid(Mass_kg = seq_range(Mass_kg, n = 183), 
            dispersal_year = seq(min(dispersal_year), max(dispersal_year), by = 7)) %>% 
  add_epred_draws(trophic_dispersal_migrators)

pred_mod_trophic_migration_migrators <- traits_migrators %>%
  group_by(diet_broadest_cat) %>%
  data_grid(Mass_kg = seq_range(Mass_kg, n = 183), 
            migration_year = seq(min(migration_year, na.rm = TRUE), max(migration_year, na.rm=TRUE), by = 7)) %>% 
  add_epred_draws(trophic_migration_migrators)

trophic_foraging_plot_migrators <- ggplot() +
  geom_point(data = traits_migrators, aes(x = Mass_kg, y = log10(hr.radius), color = ordered(diet_broadest_cat))) +
  stat_lineribbon(data = pred_mod_trophic_foraging_migrators, 
                  aes(y = log10(.epred), x = Mass_kg, color = ordered(diet_broadest_cat), fill = ordered(diet_broadest_cat)), 
                  .width = c(.95), alpha = 0.4) +
  scale_color_manual(values = diet_cols, name = "Trophic Guild")+
  scale_fill_manual(values = diet_cols, name = "Trophic Guild") +
  scale_x_log10(breaks = scales::trans_breaks("log10", function(x) 10^x),
                labels = scales::trans_format("log10", scales::math_format(10^.x)))+
  scale_y_continuous(labels = scales::math_format())+
  cowplot::theme_cowplot()+
  xlab("Body mass (kg)")+
  ylab("Foraging radius (km)")

trophic_dispersal_plot_migrators <- ggplot()+
  geom_point(data = traits_migrators, aes(x = Mass_kg, y = log10(dispersal_km), 
                                color = ordered(diet_broadest_cat))) +
  stat_lineribbon(data = pred_mod_trophic_dispersal_migrators,
                  aes(y = log10(.epred), x = Mass_kg,
                      color = ordered(diet_broadest_cat), 
                      fill = ordered(diet_broadest_cat)),
                  .width = 0.95, alpha = 0.4)+
  scale_color_manual(values = diet_cols, name = "Trophic Guild")+
  scale_fill_manual(values = diet_cols, name = "Trophic Guild") +
  scale_x_log10(breaks = scales::trans_breaks("log10", function(x) 10^x),
                labels = scales::trans_format("log10", scales::math_format(10^.x)))+
  scale_y_continuous(labels = scales::math_format())+
  cowplot::theme_cowplot()+
  xlab("Body mass (kg)")+
  ylab("Dispersal distance (km)")

trophic_migration_plot_migrators <- ggplot() +
  geom_point(data = traits_migrators, aes(x = Mass_kg, y = log10(Migration_km), color = ordered(diet_broadest_cat))) +
stat_lineribbon(data = pred_mod_trophic_migration_migrators[(pred_mod_trophic_migration_migrators$.epred>0),],
                  aes(y = log10(.epred), x = Mass_kg, 
                      color = ordered(diet_broadest_cat), 
                      fill = ordered(diet_broadest_cat)), 
                  .width = c(.95), alpha = 0.4) +
  scale_color_manual(values = diet_cols, name = "Trophic Guild")+
  scale_fill_manual(values = diet_cols, name = "Trophic Guild") + 
  scale_x_log10(breaks = scales::trans_breaks("log10", function(x) 10^x),
                labels = scales::trans_format("log10", scales::math_format(10^.x)))+
  # scale_y_log10(breaks = scales::trans_breaks("log10", function(x) 10^x),
  #               labels = scales::trans_format("log10", scales::math_format(10^.x)))+
  scale_y_continuous(labels = scales::math_format())+
  cowplot::theme_cowplot()+
  xlab("Body mass (kg)")+
  ylab("Migration distance (km)")
```

```{r render S2 figs}
gA <- ggplotGrob(class_foraging_plot_migrators)
gB <- ggplotGrob(class_dispersal_plot_migrators)
gC <- ggplotGrob(class_migration_plot_migrators)
gD <- ggplotGrob(trophic_foraging_plot_migrators)
gE <- ggplotGrob(trophic_dispersal_plot_migrators)
gF <- ggplotGrob(trophic_migration_plot_migrators)

gA$widths <- gC$widths
gB$widths <- gC$widths
gD$widths <- gC$widths
gE$widths <- gC$widths
gF$widths <- gC$widths


while (!is.null(dev.list()))  dev.off()
jpeg(filename = "../figures/FigS2_migrators_07Aug2023.jpeg", width = 10, height = 10, units = 'in', res = 1000)
grid.arrange(gA, gD, gB, gE,gC, gF, nrow = 3, padding = 2)
dev.off()
```

Non-migrators only

```{r Fig S3 non-migrators only}

traits_nonmigrators <- traits %>% filter(Migration_km == 0)

## class
class_dispersal_nonmigrators <- readRDS("../mods/class.disp_withyear_nonmigrators.rds")
class_foraging_nonmigrators <- readRDS("../mods/class.for_withyear_nonmigrators.rds")


pred_mod_class_dispersal_nonmigrators <- traits_nonmigrators %>%
  group_by(class) %>%
  data_grid(Mass_kg = seq_range(Mass_kg, n = 137), 
            dispersal_year = seq(min(dispersal_year), max(dispersal_year), by = 7)) %>% 
  add_epred_draws(class_dispersal_nonmigrators)

pred_mod_class_foraging_nonmigrators <- traits_nonmigrators %>%
  group_by(class) %>%
  data_grid(Mass_kg = seq_range(Mass_kg, n = 137), 
            foraging_year = seq(min(foraging_year, na.rm = TRUE), max(foraging_year, na.rm = TRUE), by = 7)) %>% 
  add_epred_draws(class_foraging_nonmigrators)

class_foraging_plot_nonmigrators <- ggplot() +
  geom_point(data = traits_nonmigrators, aes(x = Mass_kg, y = log10(hr.radius), color = ordered(class))) +
  stat_lineribbon(data = pred_mod_class_foraging_nonmigrators, 
                  aes(y = log10(.epred), x = Mass_kg, color = ordered(class), fill = ordered(class)), 
                  .width = c(.95), alpha = 0.4) +
  scale_color_manual(values = class_cols, name = "Class")+
  scale_fill_manual(values = class_cols, name = "Class") + 
  scale_x_log10(breaks = scales::trans_breaks("log10", function(x) 10^x),
                labels = scales::trans_format("log10", scales::math_format(10^.x)))+
  scale_y_continuous(labels = scales::math_format())+
  cowplot::theme_cowplot()+
  xlab("Body mass (kg)")+
  ylab("Foraging radius (km)")


class_dispersal_plot_nonmigrators <- ggplot()+
  geom_point(data = traits_nonmigrators, aes(x = Mass_kg, y = log10(dispersal_km), color = ordered(class))) +
  stat_lineribbon(data =pred_mod_class_dispersal_nonmigrators,
                  aes(y = log10(.epred), x = Mass_kg,
                      color = ordered(class), fill = ordered(class)),
                  .width = 0.95, alpha = 0.4)+
  scale_color_manual(values = class_cols, name = "Class")+
  scale_fill_manual(values = class_cols, name = "Class") + 
  scale_x_log10(breaks = scales::trans_breaks("log10", function(x) 10^x),
                labels = scales::trans_format("log10", scales::math_format(10^.x)))+
  scale_y_continuous(labels = scales::math_format())+
  cowplot::theme_cowplot()+
  xlab("Body mass (kg)")+
  ylab("Dispersal distance (km)")


## trophic guild
traits$diet_broadest_cat <- as.factor(traits$diet_broadest_cat)

trophic_foraging_nonmigrators <- readRDS("../mods/trophic.for_withyear_nonmigrators.rds")
trophic_dispersal_nonmigrators <- readRDS("../mods/trophic.disp_withyear_nonmigrators.rds")

pred_mod_trophic_foraging_nonmigrators <- traits_nonmigrators %>%
  group_by(diet_broadest_cat) %>%
  data_grid(Mass_kg = seq_range(Mass_kg, n = 137), 
            foraging_year = seq(min(foraging_year, na.rm = TRUE), max(foraging_year, na.rm = TRUE), by = 7)) %>% 
  add_epred_draws(trophic_foraging_nonmigrators)

pred_mod_trophic_dispersal_nonmigrators <- traits_nonmigrators %>%
  group_by(diet_broadest_cat) %>%
  data_grid(Mass_kg = seq_range(Mass_kg, n = 137), 
            dispersal_year = seq(min(dispersal_year), max(dispersal_year), by = 7)) %>% 
  add_epred_draws(trophic_dispersal_nonmigrators)


trophic_foraging_plot_nonmigrators <- ggplot() +
  geom_point(data = traits_nonmigrators, aes(x = Mass_kg, y = log10(hr.radius), color = ordered(diet_broadest_cat))) +
  stat_lineribbon(data = pred_mod_trophic_foraging_nonmigrators, 
                  aes(y = log10(.epred), x = Mass_kg, color = ordered(diet_broadest_cat), fill = ordered(diet_broadest_cat)), 
                  .width = c(.95), alpha = 0.4) +
  scale_color_manual(values = diet_cols, name = "Trophic Guild")+
  scale_fill_manual(values = diet_cols, name = "Trophic Guild") +
  scale_x_log10(breaks = scales::trans_breaks("log10", function(x) 10^x),
                labels = scales::trans_format("log10", scales::math_format(10^.x)))+
  scale_y_continuous(labels = scales::math_format())+
  cowplot::theme_cowplot()+
  xlab("Body mass (kg)")+
  ylab("Foraging radius (km)")

trophic_dispersal_plot_nonmigrators <- ggplot()+
  geom_point(data = traits_nonmigrators, aes(x = Mass_kg, y = log10(dispersal_km), 
                                color = ordered(diet_broadest_cat))) +
  stat_lineribbon(data = pred_mod_trophic_dispersal_nonmigrators,
                  aes(y = log10(.epred), x = Mass_kg,
                      color = ordered(diet_broadest_cat), 
                      fill = ordered(diet_broadest_cat)),
                  .width = 0.95, alpha = 0.4)+
  scale_color_manual(values = diet_cols, name = "Trophic Guild")+
  scale_fill_manual(values = diet_cols, name = "Trophic Guild") +
  scale_x_log10(breaks = scales::trans_breaks("log10", function(x) 10^x),
                labels = scales::trans_format("log10", scales::math_format(10^.x)))+
  scale_y_continuous(labels = scales::math_format())+
  cowplot::theme_cowplot()+
  xlab("Body mass (kg)")+
  ylab("Dispersal distance (km)")

gA <- ggplotGrob(class_foraging_plot_nonmigrators)
gB <- ggplotGrob(class_dispersal_plot_nonmigrators)
gC <- ggplotGrob(trophic_foraging_plot_nonmigrators)
gD <- ggplotGrob(trophic_dispersal_plot_nonmigrators)

gA$widths <- gB$widths
gC$widths <- gB$widths
gD$widths <- gB$widths

while (!is.null(dev.list()))  dev.off()
jpeg(filename = "../figures/FigS3_nonmigrators_03May2023.jpeg", width = 10, height = 7, units = 'in', res = 1000)
grid.arrange(gA, gC, gB, gD, nrow = 2, padding = 2)
dev.off()
```


Class posterior means, generated for Figure S4

```{r Fig S4 - class posterior means}
class_dispersal <- readRDS("../mods/class.disp_withyear.rds")
class_foraging <- readRDS("../mods/class.for_withyear.rds")
class_migration <- readRDS("../mods/class.mig_withyear.rds")


grid_foraging <-  traits %>%
  group_by(class) %>%
  data_grid(Mass_kg = seq_range(Mass_kg, n = 320), 
            foraging_year = seq(min(foraging_year, na.rm = TRUE), max(foraging_year, na.rm = TRUE), by = 7))

grid_dispersal <-  traits %>%
  group_by(class) %>%
  data_grid(Mass_kg = seq_range(Mass_kg, n = 320), 
            dispersal_year = seq(min(dispersal_year, na.rm = TRUE), max(dispersal_year, na.rm = TRUE), by = 7))

grid_migration <-  traits %>%
  group_by(class) %>%
  data_grid(Mass_kg = seq_range(Mass_kg, n = 320), 
            migration_year = seq(min(migration_year, na.rm = TRUE), max(migration_year, na.rm = TRUE), by = 7))
# foraging
means_foraging <-  grid_foraging %>%
  add_epred_draws(class_foraging)

preds_foraging <-  grid_foraging %>%
  add_predicted_draws(class_foraging)

#dispersal
means_dispersal <-  grid_dispersal %>%
  add_epred_draws(class_dispersal)

preds_dispersal <-  grid_dispersal %>%
  add_predicted_draws(class_dispersal)

#migration
means_migration <-  grid_migration %>%
  add_epred_draws(class_migration)

preds_migration <-  grid_migration %>%
  add_predicted_draws(class_migration)


class_for <- traits %>%
  ggplot(aes(y = class, x = log10(hr.radius))) +
  stat_interval(aes(x = log10(.prediction)), data = preds_dispersal[(preds_dispersal$.prediction>0),]) +
  stat_pointinterval(aes(x = log10(.epred)), data = means_foraging[(means_foraging$.epred>0),], .width = c(.66, .95), position = position_nudge(y = -0.3)) +
  geom_point() +
  scale_color_brewer()+
  cowplot::theme_cowplot()+
  labs(y = "Taxonomic class")

class_disp <- traits %>%
  ggplot(aes(y = class, x = log10(dispersal_km))) +
  stat_interval(aes(x = log10(.prediction)), 
                data = preds_dispersal[(preds_dispersal$.prediction>0),]) +
  stat_pointinterval(aes(x = log10(.epred)), 
                     data = means_dispersal[(means_dispersal$.epred>0),], 
                     .width = c(.66, .95), position = position_nudge(y = -0.3)) +
  geom_point() +
  scale_color_brewer()+
  cowplot::theme_cowplot()+
  labs(y = "Taxonomic class")

class_mig <- traits %>%
  ggplot(aes(y = class, x = log10(Migration_km))) +
  stat_interval(aes(x = log10(.prediction)), 
                data = preds_migration[(preds_migration$.prediction>0),]) +
  stat_pointinterval(aes(x = log10(.epred)), 
                     data = means_migration[(means_migration$.epred>0),], 
                     .width = c(.66, .95), position = position_nudge(y = -0.3)) +
  geom_point() +
  scale_color_brewer()+
  cowplot::theme_cowplot()+
  labs(y = "Taxonomic class")



class_figs_posterior_means <- ggarrange(class_for, class_disp, class_mig, nrow = 1, 
                        common.legend = T, legend = "right", labels = c("a", "b", "c"))


while (!is.null(dev.list()))  dev.off()
jpeg(filename = "../figures/class_figs_posterior_means_May2023.jpeg", width = 12, height = 3, units = 'in', res = 1000)
class_figs_posterior_means
dev.off()
```


Trophic guild posterior means, generated for Figure S4

```{r Fig S4 - trophic guild posterior means}
# trophic_foraging <- readRDS("../mods/trophic.for_withyear.rds")
# trophic_dispersal <- readRDS("../mods/trophic.disp_withyear.rds")
# trophic_migration <- readRDS("../mods/trophic.mig_withyear.rds")

grid_foraging_trophic <-  traits %>%
  group_by(diet_broadest_cat) %>%
  data_grid(Mass_kg = seq_range(Mass_kg, n = 320), 
            foraging_year = seq(min(foraging_year, na.rm = TRUE), max(foraging_year, na.rm = TRUE), by = 7))

grid_dispersal_trophic <-  traits %>%
  group_by(diet_broadest_cat) %>%
  data_grid(Mass_kg = seq_range(Mass_kg, n = 320), 
            dispersal_year = seq(min(dispersal_year, na.rm = TRUE), max(dispersal_year, na.rm = TRUE), by = 7))

grid_migration_trophic <-  traits %>%
  group_by(diet_broadest_cat) %>%
  data_grid(Mass_kg = seq_range(Mass_kg, n = 320), 
            migration_year = seq(min(migration_year, na.rm = TRUE), max(migration_year, na.rm = TRUE), by = 7))
# foraging
means_foraging_trophic <-  grid_foraging_trophic %>%
  add_epred_draws(trophic_foraging)

preds_foraging_trophic <-  grid_foraging_trophic %>%
  add_predicted_draws(trophic_foraging)

#dispersal
means_dispersal_trophic <-  grid_dispersal_trophic %>%
  add_epred_draws(trophic_dispersal)

preds_dispersal_trophic <-  grid_dispersal_trophic %>%
  add_predicted_draws(trophic_dispersal)

#migration
means_migration_trophic <-  grid_migration_trophic %>%
  add_epred_draws(trophic_migration)

preds_migration_trophic <-  grid_migration_trophic %>%
  add_predicted_draws(trophic_migration)


trophic_for <- traits %>%
  ggplot(aes(y = diet_broadest_cat, x = log10(hr.radius))) +
  stat_interval(aes(x = log10(.prediction)), 
                data = preds_foraging_trophic[(preds_foraging_trophic$.prediction>0),]) +
  stat_pointinterval(aes(x = log10(.epred)), data = means_foraging_trophic[(means_foraging_trophic$.epred>0),], .width = c(.66, .95), position = position_nudge(y = -0.3)) +
  geom_point() +
  scale_color_brewer()+
  cowplot::theme_cowplot()+
  labs(y = "Taxonomic class")

trophic_disp <- traits %>%
  ggplot(aes(y = diet_broadest_cat, x = log10(dispersal_km))) +
  stat_interval(aes(x = log10(.prediction)), 
                data = preds_dispersal_trophic[(preds_dispersal_trophic$.prediction>0),]) +
  stat_pointinterval(aes(x = log10(.epred)), data = means_dispersal_trophic[(means_dispersal_trophic$.epred>0),],
                     .width = c(.66, .95), position = position_nudge(y = -0.3)) +
  geom_point() +
  scale_color_brewer()+
  cowplot::theme_cowplot()+
  labs(y = "Taxonomic class")

trophic_mig <- traits %>%
  ggplot(aes(y = diet_broadest_cat, x = log10(Migration_km))) +
  stat_interval(aes(x = log10(.prediction)), 
                data = preds_migration_trophic[(preds_migration_trophic$.prediction>0),]) +
  stat_pointinterval(aes(x = log10(.epred)), 
                     data = means_migration_trophic[(means_migration_trophic$.epred>0),], 
                     .width = c(.66, .95), position = position_nudge(y = -0.3)) +
  geom_point() +
  scale_color_brewer()+
  cowplot::theme_cowplot()+
  labs(y = "Taxonomic class")



trophic_figs_posterior_means <- ggarrange(trophic_for, trophic_disp, trophic_mig, nrow = 1, 
                        common.legend = T, legend = "right", labels = c("d", "e", "f"))


while (!is.null(dev.list()))  dev.off()
jpeg(filename = "../figures/trophic_figs_posteriormeans.jpeg", width = 12, height = 3, units = 'in', res = 1000)
trophic_figs_posterior_means
dev.off()


```

Posterior mean slope distributions, as per reviewer suggestion from EcoLetts, grouped by class.

```{r Fig S5 posterior mean slopes by class}
# class_dispersal <- readRDS("../mods/class.disp_withyear.rds")
# class_foraging <- readRDS("../mods/class.for_withyear.rds")
# class_migration <- readRDS("../mods/class.mig_withyear.rds")

# foraging
coef_no_summary_for <- coef(class_foraging, summary = FALSE)
coef_no_summary_for$class %>% dimnames()
# 1 = iteration, #2 class #3 slopes 
coef_no_summary_for$class[,,] %>% dim()
apply(coef_no_summary_for$class[,,], c(2,3), FUN = function(x) c(mean(x), quantile(x, c(0.025, 0.5,0.975))))

## posterior distribution
class_foraging_slope <- coef_no_summary_for$class[,,2] %>% 
  as.data.frame() %>% 
  pivot_longer(everything(),names_to = "slope", values_to = "posterior_samples") %>% 
  ggplot(aes(x = posterior_samples)) +
  facet_wrap(~slope, ncol = 1) +
  geom_histogram(stat = "bin", bins = 100) +
  geom_vline(xintercept = 0, colour = 'red') +
  theme_cowplot()+
  theme(strip.background = element_rect(fill = "white"))+
  xlab("Posterior mean slope")


# dispersal
coef_no_summary_disp <- coef(class_dispersal, summary = FALSE)
coef_no_summary_disp$class %>% dimnames()
# 1 = iteration, #2 class #3 slopes 
coef_no_summary_disp$class[,,] %>% dim()
apply(coef_no_summary_disp$class[,,], c(2,3), FUN = function(x) c(mean(x), quantile(x, c(0.025, 0.5,0.975))))

class_dispersal_slope <- coef_no_summary_disp$class[,,2] %>% 
  as.data.frame() %>% 
  pivot_longer(everything(),names_to = "slope", values_to = "posterior_samples") %>% 
  ggplot(aes(x = posterior_samples)) +
  facet_wrap(~slope, ncol = 1) +
  geom_histogram(stat = "bin", bins = 100) +
  geom_vline(xintercept = 0, colour = 'red') +
  theme_cowplot()+
  theme(strip.background = element_rect(fill = "white"))+
  xlab("Posterior mean slope")


# migration
coef_no_summary_mig <- coef(class_migration, summary = FALSE)
coef_no_summary_mig$class %>% dimnames()
# 1 = iteration, #2 class #3 slopes 
coef_no_summary_mig$class[,,] %>% dim()
apply(coef_no_summary_mig$class[,,], c(2,3), FUN = function(x) c(mean(x), quantile(x, c(0.025, 0.5,0.975))))

class_migration_slope <- coef_no_summary_mig$class[,,2] %>% 
  as.data.frame() %>% 
  pivot_longer(everything(),names_to = "slope", values_to = "posterior_samples") %>% 
  ggplot(aes(x = posterior_samples)) +
  facet_wrap(~slope, ncol = 1) +
  geom_histogram(stat = "bin", bins = 100) +
  geom_vline(xintercept = 0, colour = 'red') +
  theme_cowplot()+
  theme(strip.background = element_rect(fill = "white"))+
  xlab("Posterior mean slope")



class_post_mean_slopes <- ggarrange(class_foraging_slope, class_dispersal_slope, class_migration_slope, 
          labels = c("Foraging", "Dispersal", "Migration"),
          nrow = 1)

jpeg(filename = "../figures/class_posterior_means_slopes.jpeg", width = 9, height = 7, units = 'in', res = 1000)
class_post_mean_slopes
dev.off()
```

Posterior mean slope distributions, as per reviewer suggestion from EcoLetts, grouped by trophic guild.


```{r Fig S6 posterior mean slopes by trophic guild}
# trophic_foraging <- readRDS("../mods/trophic.for_withyear.rds")
# trophic_dispersal <- readRDS("../mods/trophic.disp_withyear.rds")
# trophic_migration <- readRDS("../mods/trophic.mig_withyear.rds")

# foraging
coef_no_summary_for_trophic <- coef(trophic_foraging, summary = FALSE)
coef_no_summary_for_trophic$diet_broadest_cat %>% dimnames()
# 1 = iteration, #2 diet_broadest_cat #3 slopes 
coef_no_summary_for_trophic$diet_broadest_cat[,,] %>% dim()
apply(coef_no_summary_for_trophic$diet_broadest_cat[,,], c(2,3), FUN = function(x) c(mean(x), quantile(x, c(0.025, 0.5,0.975))))

## posterior distribution
trophic_foraging_slope <- coef_no_summary_for_trophic$diet_broadest_cat[,,2] %>% 
  as.data.frame() %>% 
  pivot_longer(everything(),names_to = "slope", values_to = "posterior_samples") %>% 
  ggplot(aes(x = posterior_samples)) +
  facet_wrap(~slope, ncol = 1) +
  geom_histogram(stat = "bin", bins = 100) +
  geom_vline(xintercept = 0, colour = 'red') +
  theme_cowplot()+
  theme(strip.background = element_rect(fill = "white"))+
  xlab("Posterior mean slope")


# dispersal
coef_no_summary_disp_trophic <- coef(trophic_dispersal, summary = FALSE)
coef_no_summary_disp_trophic$diet_broadest_cat %>% dimnames()
# 1 = iteration, #2 diet_broadest_cat #3 slopes 
coef_no_summary_disp_trophic$diet_broadest_cat[,,] %>% dim()
apply(coef_no_summary_disp_trophic$diet_broadest_cat[,,], c(2,3), FUN = function(x) c(mean(x), quantile(x, c(0.025, 0.5,0.975))))

trophic_dispersal_slope <- coef_no_summary_disp_trophic$diet_broadest_cat[,,2] %>% 
  as.data.frame() %>% 
  pivot_longer(everything(),names_to = "slope", values_to = "posterior_samples") %>% 
  ggplot(aes(x = posterior_samples)) +
  facet_wrap(~slope, ncol = 1) +
  geom_histogram(stat = "bin", bins = 100) +
  geom_vline(xintercept = 0, colour = 'red') +
  theme_cowplot()+
  theme(strip.background = element_rect(fill = "white"))+
  xlab("Posterior mean slope")


# migration
coef_no_summary_mig_trophic <- coef(trophic_migration, summary = FALSE)
coef_no_summary_mig_trophic$diet_broadest_cat %>% dimnames()
# 1 = iteration, #2 diet_broadest_cat #3 slopes 
coef_no_summary_mig_trophic$diet_broadest_cat[,,] %>% dim()
apply(coef_no_summary_mig_trophic$diet_broadest_cat[,,], c(2,3), FUN = function(x) c(mean(x), quantile(x, c(0.025, 0.5,0.975))))

trophic_migration_slope <- coef_no_summary_mig_trophic$diet_broadest_cat[,,2] %>% 
  as.data.frame() %>% 
  pivot_longer(everything(),names_to = "slope", values_to = "posterior_samples") %>% 
  ggplot(aes(x = posterior_samples)) +
  facet_wrap(~slope, ncol = 1) +
  geom_histogram(stat = "bin", bins = 100) +
  geom_vline(xintercept = 0, colour = 'red') +
  theme_cowplot()+
  theme(strip.background = element_rect(fill = "white"))+
  xlab("Posterior mean slope")


trophic_post_mean_slopes <- ggarrange(trophic_foraging_slope, trophic_dispersal_slope, trophic_migration_slope, 
          labels = c("Foraging", "Dispersal", "Migration"),
          nrow = 1)

jpeg(filename = "../figures/trophic_posterior_means_slopes.jpeg", width = 9, height = 7, units = 'in', res = 1000)
trophic_post_mean_slopes
dev.off()


```



Code to create covariance matrices and contingency tables, presented in Fig 3c and Tables S1, S3-S4
```{r Covariance matrix and contingency table S1}
class_diet <- table(traits$class, traits$diet_broadest_cat)
chisq.test(class_diet, simulate.p.value = TRUE, B = 10000)



library(stats)
library(gplots)

name <- traits$scientific_name
home.range <- log10(traits$hr.radius)
dispersal <- log10(traits$dispersal_km)
migration_km1 <- log10(traits$Migration_km+1)
migration <- log10(traits$Migration_km)
mass <- log10(traits$Mass_kg)

cov.mat <- data.frame(home.range, dispersal, migration_km1, mass)
matrix <- cov(cov.mat)
diag(matrix) <- NA
View(matrix)

# remove cases where migration is zero, matrix only for migrators
cov.mat_migrators <- data.frame(home.range, dispersal, migration, mass)
cov.mat2 <- cov.mat_migrators[-c(which(cov.mat_migrators$migration == "-Inf")),]

matrix2 <- cov(cov.mat2, use = "pairwise.complete.obs") 
diag(matrix2) <- NA
View(matrix2)


### non-migrators only ###
cov.mat.nonmig <- cov.mat_migrators %>% filter(migration == "-Inf")
# remove migration column
cov.mat.nonmig <- cov.mat.nonmig %>% select(-migration)

matrix.nonmig <- cov(cov.mat.nonmig, use = "pairwise.complete.obs") #go with this one
# diag(matrix2) <- NA
diag(matrix.nonmig) <- NA
View(matrix.nonmig)
```
